Explicetelly pass evidence of extension to ExtendedRpcOptions
Previous approach with writting a typeclass `ExtendingOptions o os`
which proves that `RpcConstraints os r` is not smaller than `RpcConstraints o r`
wasn't actually comprehensive, and required `IncoherentInstances`.

This commit makes user of `ExtendedRpcOptions` to provide an evidence of that
inequality on constraints manually.
In most cases she has to do it for specific
`o` and `os`, thus providing an evidence is as easy as passing `Evi Dict`.
And even if one needs a polymorphic function on those `o` and `os`, she can
just carry the evidence around.